// server/src/orders/enum/status.enum.ts
export enum OrderStatus {
  PENDING = 'pending',
  IN_PREPARATION = 'in-preparation',
  READY_FOR_PICKUP = 'ready-for-pickup',
  COMPLETED = 'completed',
  CANCELLED = 'cancelled',
}

export enum OrderStatusId {
  PENDING = 1,
  IN_PREPARATION = 2,
  READY_FOR_PICKUP = 3,
  COMPLETED = 4,
  CANCELLED = 5,
}